Revert "validate instance type in all operations"
This reverts commit 56ef82ddc8e2d073464e8fb567fc68a4337294da.

+++ cfn-resources/cloud-backup-snapshot/cmd/resource/resource.go
@@ -43,24 +43,10 @@ func setup() {
 	util.SetupLogger("mongodb-atlas-cloud-backup-snapshot")
 }
 
-func validateModel(fields []string, model *Model) *handler.ProgressEvent {
-	if pe := validator.ValidateModel(fields, model); pe != nil {
-		return pe
-	}
-
-	if *model.InstanceType != clusterInstanceType {
-		pe := progressevent.GetFailedEventByCode(fmt.Sprintf("InstanceType must be %s", clusterInstanceType),
-			string(types.HandlerErrorCodeInvalidRequest))
-		return &pe
-	}
-
-	return nil
-}
-
 func Create(req handler.Request, prevModel *Model, currentModel *Model) (handler.ProgressEvent, error) {
 	setup()
 	util.SetDefaultProfileIfNotDefined(&currentModel.Profile)
-	if err := validateModel(CreateRequiredFields, currentModel); err != nil {
+	if err := validator.ValidateModel(CreateRequiredFields, currentModel); err != nil {
 		return *err, nil
 	}
 
@@ -75,33 +61,36 @@ func Create(req handler.Request, prevModel *Model, currentModel *Model) (handler
 		return validateProgress(client, currentModel, "completed")
 	}
 
-	params := admin20231115002.DiskBackupOnDemandSnapshotRequest{
-		Description:     currentModel.Description,
-		RetentionInDays: currentModel.RetentionInDays,
-	}
-	snapshot, resp, err := client.Atlas20231115002.CloudBackupsApi.TakeSnapshot(context.Background(), *currentModel.ProjectId, *currentModel.InstanceName, &params).Execute()
-	if err != nil {
-		return progressevent.GetFailedEventByResponse(err.Error(), resp), nil
-	}
+	if *currentModel.InstanceType == clusterInstanceType {
+		params := admin20231115002.DiskBackupOnDemandSnapshotRequest{
+			Description:     currentModel.Description,
+			RetentionInDays: currentModel.RetentionInDays,
+		}
+		snapshot, resp, err := client.Atlas20231115002.CloudBackupsApi.TakeSnapshot(context.Background(), *currentModel.ProjectId, *currentModel.InstanceName, &params).Execute()
+		if err != nil {
+			return progressevent.GetFailedEventByResponse(err.Error(), resp), nil
+		}
 
-	currentModel.SnapshotId = snapshot.Id
+		currentModel.SnapshotId = snapshot.Id
 
-	return handler.ProgressEvent{
-		OperationStatus:      handler.InProgress,
-		Message:              fmt.Sprintf("Create cloud provider snapshots : %s", *snapshot.Status),
-		ResourceModel:        currentModel,
-		CallbackDelaySeconds: 65,
-		CallbackContext: map[string]interface{}{
-			"status":      snapshot.Status,
-			"snapshot_id": snapshot.Id,
-		},
-	}, nil
+		return handler.ProgressEvent{
+			OperationStatus:      handler.InProgress,
+			Message:              fmt.Sprintf("Create cloud provider snapshots : %s", *snapshot.Status),
+			ResourceModel:        currentModel,
+			CallbackDelaySeconds: 65,
+			CallbackContext: map[string]interface{}{
+				"status":      snapshot.Status,
+				"snapshot_id": snapshot.Id,
+			},
+		}, nil
+	}
+	return handler.ProgressEvent{}, fmt.Errorf("unsupported InstanceType: %s", *currentModel.InstanceType)
 }
 
 func Read(req handler.Request, prevModel *Model, currentModel *Model) (handler.ProgressEvent, error) {
 	setup()
 	util.SetDefaultProfileIfNotDefined(&currentModel.Profile)
-	if err := validateModel(ReadRequiredFields, currentModel); err != nil {
+	if err := validator.ValidateModel(ReadRequiredFields, currentModel); err != nil {
 		return *err, nil
 	}
 
@@ -114,11 +103,13 @@ func Read(req handler.Request, prevModel *Model, currentModel *Model) (handler.P
 		return *pe, nil
 	}
 
-	server, resp, err := client.Atlas20231115002.CloudBackupsApi.GetReplicaSetBackup(context.Background(), *currentModel.ProjectId, *currentModel.InstanceName, *currentModel.SnapshotId).Execute()
-	if err != nil {
-		return progressevent.GetFailedEventByResponse(err.Error(), resp), nil
+	if *currentModel.InstanceType == clusterInstanceType {
+		server, resp, err := client.Atlas20231115002.CloudBackupsApi.GetReplicaSetBackup(context.Background(), *currentModel.ProjectId, *currentModel.InstanceName, *currentModel.SnapshotId).Execute()
+		if err != nil {
+			return progressevent.GetFailedEventByResponse(err.Error(), resp), nil
+		}
+		currentModel.updateModelServer(server)
 	}
-	currentModel.updateModelServer(server)
 
 	return handler.ProgressEvent{
 		OperationStatus: handler.Success,
@@ -134,7 +125,7 @@ func Update(req handler.Request, prevModel *Model, currentModel *Model) (handler
 func Delete(req handler.Request, prevModel *Model, currentModel *Model) (handler.ProgressEvent, error) {
 	setup()
 	util.SetDefaultProfileIfNotDefined(&currentModel.Profile)
-	if err := validateModel(DeleteRequiredFields, currentModel); err != nil {
+	if err := validator.ValidateModel(DeleteRequiredFields, currentModel); err != nil {
 		return *err, nil
 	}
 
@@ -147,9 +138,11 @@ func Delete(req handler.Request, prevModel *Model, currentModel *Model) (handler
 		return *pe, nil
 	}
 
-	_, resp, err := client.Atlas20231115002.CloudBackupsApi.DeleteReplicaSetBackup(context.Background(), *currentModel.ProjectId, *currentModel.InstanceName, *currentModel.SnapshotId).Execute()
-	if err != nil {
-		return progressevent.GetFailedEventByResponse(err.Error(), resp), nil
+	if *currentModel.InstanceType == clusterInstanceType {
+		_, resp, err := client.Atlas20231115002.CloudBackupsApi.DeleteReplicaSetBackup(context.Background(), *currentModel.ProjectId, *currentModel.InstanceName, *currentModel.SnapshotId).Execute()
+		if err != nil {
+			return progressevent.GetFailedEventByResponse(err.Error(), resp), nil
+		}
 	}
 
 	return handler.ProgressEvent{
@@ -161,7 +154,7 @@ func Delete(req handler.Request, prevModel *Model, currentModel *Model) (handler
 func List(req handler.Request, prevModel *Model, currentModel *Model) (handler.ProgressEvent, error) {
 	setup()
 	util.SetDefaultProfileIfNotDefined(&currentModel.Profile)
-	if err := validateModel(ListRequiredFields, currentModel); err != nil {
+	if err := validator.ValidateModel(ListRequiredFields, currentModel); err != nil {
 		return *err, nil
 	}
 
@@ -172,19 +165,21 @@ func List(req handler.Request, prevModel *Model, currentModel *Model) (handler.P
 
 	models := make([]interface{}, 0)
 
-	server, resp, err := client.Atlas20231115002.CloudBackupsApi.ListReplicaSetBackups(context.Background(), *currentModel.ProjectId, *currentModel.InstanceName).Execute()
-	if err != nil {
-		return progressevent.GetFailedEventByResponse(err.Error(), resp), nil
-	}
-	for i := range server.Results {
-		model := Model{
-			ProjectId:    currentModel.ProjectId,
-			Profile:      currentModel.Profile,
-			InstanceName: currentModel.InstanceName,
-			InstanceType: currentModel.InstanceType,
+	if *currentModel.InstanceType == clusterInstanceType {
+		server, resp, err := client.Atlas20231115002.CloudBackupsApi.ListReplicaSetBackups(context.Background(), *currentModel.ProjectId, *currentModel.InstanceName).Execute()
+		if err != nil {
+			return progressevent.GetFailedEventByResponse(err.Error(), resp), nil
 		}
-		model.updateModelServer(&server.Results[i])
-		models = append(models, &model)
+		for i := range server.Results {
+			model := Model{
+				ProjectId:    currentModel.ProjectId,
+				Profile:      currentModel.Profile,
+				InstanceName: currentModel.InstanceName,
+				InstanceType: currentModel.InstanceType,
+			}
+			model.updateModelServer(&server.Results[i])
+			models = append(models, &model)
+		}
 	}
 
 	return handler.ProgressEvent{
@@ -195,14 +190,16 @@ func List(req handler.Request, prevModel *Model, currentModel *Model) (handler.P
 }
 
 func validateExist(client *util.MongoDBClient, model *Model) *handler.ProgressEvent {
-	server, resp, err := client.Atlas20231115002.CloudBackupsApi.ListReplicaSetBackups(context.Background(), *model.ProjectId, *model.InstanceName).Execute()
-	if err != nil {
-		pe := progressevent.GetFailedEventByResponse(err.Error(), resp)
-		return &pe
-	}
-	for i := range server.Results {
-		if util.AreStringPtrEqual(model.SnapshotId, server.Results[i].Id) {
-			return nil
+	if *model.InstanceType == clusterInstanceType {
+		server, resp, err := client.Atlas20231115002.CloudBackupsApi.ListReplicaSetBackups(context.Background(), *model.ProjectId, *model.InstanceName).Execute()
+		if err != nil {
+			pe := progressevent.GetFailedEventByResponse(err.Error(), resp)
+			return &pe
+		}
+		for i := range server.Results {
+			if util.AreStringPtrEqual(model.SnapshotId, server.Results[i].Id) {
+				return nil
+			}
 		}
 	}
